Right... Spotted by Maksim. By the way, maybe we should start checking for an acceptable image size? If we get bogus data in the image we might end up allocating QImage with width=height=65535, which is damn big for a bogus image. (thanks Maksim)
svn path=/trunk/kdelibs/; revision=196614

+++ kimgio/krl.cpp
@@ -43,7 +43,7 @@ void kimgio_krl_read( QImageIO *iio )
 	Q_INT16 w = LOCAL_ENDIAN( buffer[ 35 ], buffer[ 34 ] );
 	Q_INT16 h = LOCAL_ENDIAN( buffer[ 37 ], buffer[ 36 ] );
 
-	Q_UINT32 samples = (int)(w*h);
+	Q_UINT32 samples = w*h;
 
 	//debug( "kimgio_krl_read: image w: %d, h: %d samples: %d",
 	//		(int)w, (int)h, samples );
